wip: download plugins before build
