+++ brack-plugin/src/plugin.rs
@@ -10,9 +10,13 @@ pub type CommandName = String;
 pub type Plugins = HashMap<ModuleName, (Plugin, PluginMetaDataMap)>;
 pub type PluginMetaDataMap = HashMap<(CommandName, Type), MetaData>;
 
-pub fn new_plugins<P: AsRef<Path>>(pathes: Vec<P>) -> Result<Plugins> {
+pub fn new_plugins<M: AsRef<str>, P: AsRef<Path>>(
+    plutin_path_map: HashMap<M, P>,
+) -> Result<Plugins> {
     let mut result = HashMap::new();
-    for path in pathes {
+    for (name, path) in plutin_path_map {
+        let name = name.as_ref().to_string();
+        let path = path.as_ref().to_path_buf();
         let mut extism_plugin = Plugin::new(fs::read(&path)?, [], true)?;
         let Json(metadatas) = extism_plugin.call::<(), Json<Vec<MetaData>>>("get_metadata", ())?;
         let mut metadata_map = HashMap::new();
@@ -22,13 +26,6 @@ pub fn new_plugins<P: AsRef<Path>>(pathes: Vec<P>) -> Result<Plugins> {
                 metadata,
             );
         }
-        let file_stem = path
-            .as_ref()
-            .file_stem()
-            .and_then(|s| s.to_str())
-            .unwrap_or_default();
-        let parts: Vec<&str> = file_stem.split('.').collect();
-        let name = parts.get(0).map_or("", |s| *s).to_string();
         result.insert(name, (extism_plugin, metadata_map));
     }
     Ok(result)
